Moved validation to checkInputs.js to reuse
